+++ engine/src/main/scala/cromwell/engine/workflow/workflowstore/workflowstore_.scala
@@ -1,6 +1,6 @@
 package cromwell.engine.workflow.workflowstore
 
-import cromwell.core.{WorkflowId, WorkflowSourceFiles}
+import cromwell.core.{WorkflowId, WorkflowSourceFilesCollection}
 import cromwell.engine.workflow.workflowstore.WorkflowStoreState.StartableState
 
 sealed trait WorkflowStoreState {def isStartable: Boolean}
@@ -12,4 +12,4 @@ object WorkflowStoreState {
   case object Restartable extends StartableState
 }
 
-final case class WorkflowToStart(id: WorkflowId, sources: WorkflowSourceFiles, state: StartableState)
+final case class WorkflowToStart(id: WorkflowId, sources: WorkflowSourceFilesCollection, state: StartableState)