+++ engine/src/main/scala/cromwell/webservice/CromwellApiHandler.scala
@@ -61,12 +61,16 @@ class CromwellApiHandler(requestHandlerActor: ActorRef) extends Actor with Workf
         case _ => RequestComplete((StatusCodes.InternalServerError, APIResponse.error(e)))
       }
 
-    case ApiHandlerWorkflowSubmit(source) => requestHandlerActor ! WorkflowStoreActor.SubmitWorkflow(source)
+    case ApiHandlerWorkflowSubmit(source) => requestHandlerActor ! WorkflowStoreActor.SubmitWorkflow(WorkflowSourceFiles(source.wdlSource,
+                                                                                                                          source.inputsJson,
+                                                                                                                          source.workflowOptionsJson))
 
     case WorkflowStoreActor.WorkflowSubmittedToStore(id) =>
       context.parent ! RequestComplete((StatusCodes.Created, WorkflowSubmitResponse(id.toString, WorkflowSubmitted.toString)))
 
-    case ApiHandlerWorkflowSubmitBatch(sources) => requestHandlerActor ! WorkflowStoreActor.BatchSubmitWorkflows(sources)
+    case ApiHandlerWorkflowSubmitBatch(sources) => requestHandlerActor !
+      WorkflowStoreActor.BatchSubmitWorkflows(sources.map(x => WorkflowSourceFiles(x.wdlSource,x.inputsJson,x.workflowOptionsJson)))
+
 
     case WorkflowStoreActor.WorkflowsBatchSubmittedToStore(ids) =>
       val responses = ids map { id => WorkflowSubmitResponse(id.toString, WorkflowSubmitted.toString) }