+++ engine/src/test/scala/cromwell/engine/workflow/WorkflowDescriptorBuilder.scala
@@ -3,7 +3,7 @@ package cromwell.engine.workflow
 import akka.actor.{ActorSystem, Props}
 import com.typesafe.config.ConfigFactory
 import cromwell.CromwellTestKitSpec
-import cromwell.core.{WorkflowId, WorkflowSourceFiles}
+import cromwell.core.{WorkflowId, WorkflowSourceFilesCollection}
 import cromwell.engine.EngineWorkflowDescriptor
 import cromwell.engine.workflow.lifecycle.MaterializeWorkflowDescriptorActor
 import cromwell.engine.workflow.lifecycle.MaterializeWorkflowDescriptorActor.{MaterializeWorkflowDescriptorCommand, MaterializeWorkflowDescriptorFailureResponse, MaterializeWorkflowDescriptorSuccessResponse, WorkflowDescriptorMaterializationResult}
@@ -15,7 +15,7 @@ trait WorkflowDescriptorBuilder {
   implicit val awaitTimeout = CromwellTestKitSpec.TimeoutDuration
   implicit val actorSystem: ActorSystem
 
-  def createMaterializedEngineWorkflowDescriptor(id: WorkflowId, workflowSources: WorkflowSourceFiles): EngineWorkflowDescriptor = {
+  def createMaterializedEngineWorkflowDescriptor(id: WorkflowId, workflowSources: WorkflowSourceFilesCollection): EngineWorkflowDescriptor = {
     import akka.pattern.ask
     implicit val timeout = akka.util.Timeout(awaitTimeout)
     implicit val ec = actorSystem.dispatcher