+++ src/main/scala/cromwell/CromwellCommandLine.scala
@@ -6,7 +6,7 @@ import better.files._
 import cats.data.Validated._
 import cats.syntax.cartesian._
 import cats.syntax.validated._
-import cromwell.core.WorkflowSourceFiles
+import cromwell.core.{WorkflowSourceFiles, WorkflowSourceFilesCollection, WorkflowSourceFilesWithImports}
 import cromwell.util.FileUtil._
 import lenthall.exception.MessageAggregation
 import cromwell.core.ErrorOr._
@@ -17,7 +17,7 @@ sealed abstract class CromwellCommandLine
 case object UsageAndExit extends CromwellCommandLine
 case object RunServer extends CromwellCommandLine
 final case class RunSingle(wdlPath: Path,
-                           sourceFiles: WorkflowSourceFiles,
+                           sourceFiles: WorkflowSourceFilesCollection,
                            inputsPath: Option[Path],
                            optionsPath: Option[Path],
                            metadataPath: Option[Path]) extends CromwellCommandLine
@@ -26,7 +26,7 @@ object CromwellCommandLine {
   def apply(args: Seq[String]): CromwellCommandLine = {
     args.headOption match {
       case Some("server") if args.size == 1 => RunServer
-      case Some("run") if args.size >= 2 && args.size <= 5 => RunSingle(args.tail)
+      case Some("run") if args.size >= 2 && args.size <= 6 => RunSingle(args.tail)
       case _ => UsageAndExit
     }
   }
@@ -38,15 +38,19 @@ object RunSingle {
     val inputsPath = argPath(args, 1, Option(".inputs"), checkDefaultExists = false)
     val optionsPath = argPath(args, 2, Option(".options"), checkDefaultExists = true)
     val metadataPath = argPath(args, 3, None)
+    val importPath = argPath(args, 4, None)
 
     val wdl = readContent("WDL file", wdlPath)
     val inputsJson = readJson("Inputs", inputsPath)
     val optionsJson = readJson("Workflow Options", optionsPath)
 
-    val sourceFiles = (wdl |@| inputsJson |@| optionsJson) map { WorkflowSourceFiles.apply }
+    val sourceFileCollection = importPath match {
+      case Some(p) => (wdl |@| inputsJson |@| optionsJson |@| validateImportsDirectory(p)) map WorkflowSourceFilesWithImports.apply
+      case None => (wdl |@| inputsJson |@| optionsJson) map WorkflowSourceFiles.apply
+    }
 
     val runSingle = for {
-      sources <- sourceFiles
+      sources <- sourceFileCollection
       _ <- writeableMetadataPath(metadataPath)
     } yield RunSingle(wdlPath, sources, inputsPath, optionsPath, metadataPath)
 
@@ -58,6 +62,29 @@ object RunSingle {
       }
     }
   }
+
+  private def validateImportsDirectory(path: Path): ErrorOr[File] = {
+
+    def unZipFile(f: File): File = {
+      val unzippedFile = f.unzip()
+      val unzippedFileContents = unzippedFile.toJava.listFiles().head
+
+      if (unzippedFileContents.isDirectory) File(unzippedFileContents.getPath)
+      else unzippedFile
+    }
+
+    val importsFile: File = File(path).extension match {
+      case Some(".zip") => unZipFile(File(path))
+      case _ => File(path)
+    }
+
+    importsFile match {
+      case file if !file.isDirectory => s"Unable to import workflows as the given path is not a directory: ${file.pathAsString}".invalidNel
+      case file if file.isDirectory && file.isEmpty => s"Unable to import workflows as the given path is an empty directory: ${file.pathAsString}".invalidNel
+      case file => file.validNel
+    }
+  }
+
 
   private def writeableMetadataPath(path: Option[Path]): ErrorOr[Unit] = {
     path match {